% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{data}
\name{Craneur}
\alias{Craneur}
\title{Craneur}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Craneur
}
\description{
R6 class for RA? creation
}
\section{Methods}{

\describe{
  \item{\code{name}}{name of the RAN}
  \item{\code{add_package}}{add a source package to the RAN}
  \item{\code{packages}}{Current packages in the RAN}
  \item{\code{paths}}{Paths to packages in the RAN}
  \item{\code{write}}{Save the RAN}
}
}

\examples{
\dontrun{
library(craneur)
colin <- Craneur$new("Colin")
colin$add_package("../craneur_0.0.0.9000.tar.gz")
colin$add_package("../jekyllthat_0.0.0.9000.tar.gz")
colin$add_package("../tidystringdist_0.1.2.tar.gz")
colin$add_package("../attempt_0.2.1.tar.gz")
colin$add_package("../rpinterest_0.4.0.tar.gz")
colin$add_package("../rgeoapi_1.2.0.tar.gz")
colin$add_package("../proustr_0.3.0.9000.tar.gz")
colin$add_package("../languagelayeR_1.2.3.tar.gz")
colin$add_package("../fryingpane_0.0.0.9000.tar.gz")
colin$add_package("../dockerfiler_0.1.1.tar.gz")
colin$add_package("../devaddins_0.0.0.9000.tar.gz")
colin$write(path = ".")}
}
\keyword{datasets}
